add start.php to store SESSION
